requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg==0.29.0
async-lru==2.0.4
Cartopy==0.22.0
foursquare==1!2020.1.30
matplotlib==3.8.3
OSMPythonTools==0.3.5
pydantic==2.5.3
pydantic-settings==2.2.1
pyowm==3.3.0
PyYAML==6.0.1
redis>=4.2.0rc1
SQLAlchemy==2.0.28